This patch adds some benchmark tests to address the performance issue
related to the context logger. The follow command be used to execute
these benchmarks from the root of the project:

    $ go test -bench=. ./api/context 2> /dev/null
    PASS
    BenchmarkStandardLog1-8           300000          4970 ns/op
    BenchmarkStandardLog10-8          200000          8962 ns/op
    BenchmarkStandardLog100-8          30000         51700 ns/op
    BenchmarkStandardLog1000-8          3000        533795 ns/op
    BenchmarkContextLogDebug1-8       100000         16598 ns/op
    BenchmarkContextLogDebug10-8       30000         57492 ns/op
    BenchmarkContextLogDebug100-8       1000       1259054 ns/op
    BenchmarkContextLogDebug1000-8        20      87484041 ns/op
    BenchmarkContextLogError1-8      1000000          1074 ns/op
    BenchmarkContextLogError10-8     1000000          1086 ns/op
    BenchmarkContextLogError100-8    1000000          1104 ns/op
    BenchmarkContextLogError1000-8   1000000          1115 ns/op
    ok      github.com/emccode/libstorage/api/context   19.034s

Additionally, this patch also provides several optimizations to the
context logger logic in order to make it more efficient and improving
overall, general performance.